feat(liability): migrate PaymentRecord.amount Float->Numeric(18,2) (#9)
Review follow-up #9 scope decision: PaymentRecord.amount migrated to Numeric;
Activity.amount stays Float.

Decision rationale:
- PaymentRecord is a payment ledger (each repayment amount) — same nature as
  WishSavingsLog (already Numeric). record_payment passes a Decimal that
  SQLAlchemy was silently coercing to Float (precision loss on payment history).
  Migrating mirrors WishSavingsLog + the liability money fields.
- Activity is a cross-entity display snapshot (asset purchase_price [Float],
  liability original_amount [Decimal], payment amount [Decimal]) with zero
  aggregation. Keeping it Float avoids a reverse Decimal->Float coerce for
  asset-origin amounts; it's not a ledger.

Backend:
- models/payment_record.py: amount Float->Numeric(18,2), Mapped[float]->Decimal.
- schemas/liability.py: new PaymentRecordResponse (amount: str via shared
  _coerce_money_str) — get_payments had no response_model; adding it makes the
  Decimal->str serialization explicit (money-as-str).
- routers/liabilities.py: get_payments gains response_model=list[PaymentRecordResponse];
  record_activity calls coerce Decimal->float (Activity stays Float).
- alembic c3d4e5f6a7b8: batch_alter payment_records.amount Float->Numeric(18,2).
- conftest.py: register PaymentRecord so test create_all builds the table.

Frontend: types/index.ts PaymentRecord.amount number->string (getPayments has
no live caller — dead API client — no display site needs Number()).

Tests: test_get_payments_returns_str_amount (Decimal->str regression);
test_export_repro.py + test_internal_repro.py (reviewer repro tests for the
export + ai_internal Decimal fixes in 12f7b948, now committed).

Verification: backend 49 passed; frontend 923/923; typecheck clean; ruff clean;
mypy net -2 errors (record_activity fix removed 2; remaining 11 are pre-existing
int/str path-param mismatches).

+++ server/apps/backend/app/models/payment_record.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Text, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Numeric, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from apps.backend.app.database import Base
@@ -12,6 +13,9 @@ class PaymentRecord(Base):
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, default=next_id)
     liability_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("liabilities.id"), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    # Money: NUMERIC(18,2) — Decimal in Python, str on the wire (money-as-str).
+    # Was Float pre-review-followup (silent precision loss on payment history);
+    # migrated to Numeric to mirror WishSavingsLog + the liability money fields.
+    amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     paid_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
